Moved `pow` into its own file
